add stage scene
